\name{RAdSense-package}
\alias{RAdSense-package}
\alias{RAdSense}
\docType{package}
\title{
\packageTitle{RAdSense}
}
\description{
\packageDescription{RAdSense}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RAdSense}
\packageIndices{RAdSense}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RAdSense}

Maintainer: \packageMaintainer{RAdSense}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
